backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
djangorestframework==3.16.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
